requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
crawl4ai>=0.7.4
openai>=1.68.0
prisma==0.15.0
pydantic>=2.10.0
python-dotenv>=1.0.1
